Assemble core infrastructure for Climate projects
Extract all of the core infrastructure components from the Climate Change API
repository and migrate it to this repository. From there, emit all of the
necessary outputs for the existing Climate Change API project to deploy itself
as a service in the core container service cluster.

In addition, this change set:

- Encapsulates ECS cluster creation with a Terraform module
- Standardizes all Terraform variables
- Simplifies dependent IAM resources
- Migrates infrastructure state to a new remote location