Remove RDS database name output
When an RDS instance is created, it automatically creates a PostgreSQL
database. Since the existing RDS instance was created for the Climate
API, that's the database name returned by this output. For other
services depending on a set of core AWS resources, retrieving this
reference, or making use of it, is only desirable when the service is
the Climate API.